refactor(audio): widen onApiResponse data type to unknown
The type for the `data` property within the `onApiResponse` callback is changed from `Record<string, unknown>` to `unknown`. This makes the component more flexible by not assuming the API response data will always be a key-value object.

BREAKING CHANGE: The `data` property in the `onApiResponse` callback is now of type `unknown`. Consumers must perform their own type checking or casting on the response data before use.